03-pdf_chatbot/pyproject.toml
[project]
name = "pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdulrehman104", email = "abdulrehman104+abdulrehman104@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.1.5",
    "langchain-qdrant>=0.2.0",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
pdf = "pdf_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
